+++ app/blueprints/__init__.py
@@ -2,3 +2,4 @@ from .charedit import *
 from .auth  import *
 from .api import *
 from .marketplace import *
+from .party import *
