+++ app/obsolete/party_edit.html
@@ -1,16 +1,18 @@
 {% extends "base.html" %} {% block content %}
 <script>
-  const party_id = `{{party.id}}`;
-  const username = `{{username}}`;
-  const url_name = `{{url_name}}`;
-  const selected_container = `{{inventory.selected_container["id"]}}`;
+  const partyID = `{{ party_id }}`;
+  let members = `{{ members }}`;
+  const joinCode = `{{ join_code }}`;
+  const userID = `{{ user_id }}`;
+  const isOwner = `{{ is_owner }}`;
+  const isSubOwner = `{{ is_subowner }}`;
+  const ownername = `{{ ownername }}`;
+  const username = `{{ username }}`;
+  const partyURL = `{{ party_url }}`;
+  const base_url = `{{ base_url }}`;
+  const data = JSON.parse(`{{ items_json|safe }}`.slice(1, -1));
+  const containersData = JSON.parse(`{{ containers_data_json|safe }}`.slice(1, -1));
 </script>
-{% if is_owner or is_subowner %}
-<script>
-  joinCode = `{{ join_code }}`;
-  const isOwner = "{{ is_owner }}";
-</script>
-{% endif %}
 <script src="{{ url_for('static', filename='src/js/party_edit.js') }}" defer type="module"></script>
 <section class="body-container">
   {% with messages = get_flashed_messages() %} {% if messages is not none and messages|length > 0 %}
@@ -24,24 +26,13 @@
     </div>
   </div>
   {% endif %} {% endwith %}
-  <div id="party-form" class="sheet box form">
+  <form id="party-form" class="sheet box" method="post">
+    {{ form.hidden_tag() }}
     <div style="display: flex; justify-content: space-between">
       <div>
-        <h1 class="title level-item justify-centered">{{ name }}</h1>
-        <p class="justify-centered"></p>
-        <p>{{description}}</p>
-      </div>
-
-      <div class="row32 center-items">
-        {% if is_owner or is_subowner %}
-        <button type="button" class="dice-button ghost40" hx-post="/party/{{username}}/{{party.id}}/save">
-          <i class="fa-solid fa-check check"></i>
-        </button>
-        <button class="dice-button ghost40" type="button" hx-post="/charedit/{{username}}/{{party.id}}/cancel">
-          <i class="fa-solid fa-xmark xmark"></i>
-        </button>
-        {% endif %}
-      </div>
+        <p class="justify-centered">{{ form.name(class='title justify-centered', id ='party-name-field') }}</p>
+        <div> {{ form.description(class='textarea') }} </div></div
+      >
     </div>
 
     <div class="party-section">
@@ -52,15 +43,28 @@
       <div id="party-members-content" class="party-section-content">
         {% if characters %} {% for character in characters %}
 
-        <div class="card character-card party-member-card" style="cursor: pointer; padding: 20px"
-          hx-get="/party/show-user/{{character.owner_username}}/{{character.url_name}}">
+        <div
+          class="card character-card"
+          style="cursor: pointer; height: auto; padding: 1em"
+          data-character-owner="{{ character.owner_username }}"
+          data-character-name="{{ character.name }}"
+          data-character-portrait-url="{{ character.image_url }}"
+          data-character-url="{{ character.url_name }}"
+          data-character-id="{{ character.id }}"
+          data-character-items="{{ character.items }}"
+          data-character-hp-max="{{ character.hp_max }}"
+        >
           <div class="card-content">
             <div class="card-image">
               <figure class="" style="display: flex; align-items: center; justify-content: center">
-                <img src="{{ character.portrait_src }}" alt="Character Portrait"
+                <img
+                  src="{{ character.portrait_src }}"
+                  alt="Character Portrait"
                   style="height: 120px; width: 120px; object-fit: cover"
-                  data-custom-image="{{ character.custom_image }}" data-image-url="{{ character.image_url }}"
-                  class="character-card-image" />
+                  data-custom-image="{{ character.custom_image }}"
+                  data-image-url="{{ character.image_url }}"
+                  class="character-card-image"
+                />
               </figure>
             </div>
             <div class="character-card-name-container">
@@ -100,6 +104,16 @@
             </div>
             <div class="level is-mobile"> </div>
           </div>
+          {% if is_owner %}
+          <footer class="card-footer edit-mode">
+            <button
+              class="card-footer-item button is-danger is-light edit-mode character-remove-button"
+              type="button"
+              id="character-remove-button"
+              >Remove
+            </button>
+          </footer>
+          {% endif %}
         </div>
 
         {% endfor %} {% else %}
@@ -113,14 +127,25 @@
       <span id="party-inventory-collapse-button" class="party-collapse-button">
         <i class="fa-solid fa-chevron-down" id="party-inventory-collapse-icon"></i>
       </span>
-      <span class="subtitle bold">Storage</span>
+      <span class="subtitle bold">Inventory</span>
       <div id="party-inventory-content" class="party-section-content">
-        <div> {% include "partial/partyview/inventory.html" %}</div>
+        <div> {% include "main/inventory.html" %}</div>
       </div>
     </div>
 
+    {% if is_owner %}
+    <div>
+      <button class="button is-danger edit-mode" type="button" id="delete-party-button" style="margin: 40px 0px">
+        Delete Party
+      </button>
+    </div>
+    {% endif %}
 
-  </div>
+    <div id="save-button-footer-wrapper" class="sticky-save-button-wrapper edit-mode">
+      <button id="save-button" class="button is-success" style="margin-right: 40px"> Save </button>
+      <button id="cancel-button" class="button" type="button"> Cancel </button>
+    </div>
+  </form>
 </section>
 
 {% endblock %}
