+++ app/static/src/scss/main.scss
@@ -706,6 +706,14 @@ button,
   width: 100%;
 }
 
+.center-items {
+  align-items: center;
+}
+
+.no-margin-top {
+  margin-top: 0px;
+}
+
 /* --- checkbox --- */
 
 .checkbox {
