+++ app/templates/partial/partyview/inventory.html
@@ -1,0 +1,34 @@
+<div class="inventory-header flex-row">
+  {% for cnt in inventory.containers %}
+  <div
+    class="inventory-container-title subtitle pointer {% if cnt.is_selected %} inventory-container-title-selected {% endif %} {% if cnt.encumbered %} red-text {% endif %}"
+    hx-get=" /charedit/inventory-select-container/{{username}}/{{url_name}}/{{cnt.id}}"
+    hx-target="#inventory-container">
+    {{cnt.title}}
+  </div>
+  {% endfor %}
+</div>
+<div style="margin-top: 0px; width: 100%">
+  {% for it in inventory.selected_container['items'] %}
+  <div class="inventory-item-container {% if it.blocker %} item-blocker-text {% endif %}">
+    {% if it.is_empty %}
+    <div class="inventory-item-text"></div>
+    {% else %}
+    <div class="inventory-item-text">
+      {{it.title | safe}}
+    </div>
+    <div class="inventory-item-button-container">
+      {% if "description" in it and it.description != "" %}
+      <div class="pointer" title="Description" onclick="KW_alert('{{it.name}}','{{it.description|eol2br|squote2js}}')">
+        <i class="item-edit-button fa-solid fa-circle-info circle-info"></i>
+      </div>
+      {% endif %} {% if it.dice|length > 0 %}
+      <div class="pointer" title="Roll dice" onclick="KW_rollDiceCallback({{it.dice}})">
+        <i class="fa-solid fa-dice dice"></i>
+      </div>
+      {% endif %}
+    </div>
+    {% endif %}
+  </div>
+  {% endfor %}
+</div>
